Added Emit object. (#141)
* Added Emit object.
This is the first version of Emit object. it's only the skeleton of the app. no methods added. Since Emit API is completely different from other APIs in AEDT, a deep custom development has to be done on emit.py further

* Added Emit object.
This is the first version of Emit object. it's only the skeleton of the app. no methods added. Since Emit API is completely different from other APIs in AEDT, a deep custom development has to be done on emit.py further