Update EDB_Data.py
@akaszynski 
I went through and used a noun for most property docstrings. However, you said that when a property has a setter, no documentation is necessary for the setter, but where it's defined in the @property, you'd document it with "Retrieve or update....." For @ properties with setters, I added the "Retrieve or update" the docstring for  the @property, which makes it a verb and not a noun. Did the "Retrieve or update ...." go somewhere else?  Also, should the @aedt_exeception entries have docstrings? I tried to find this info in the numpydoc style guide but failed. I did notice this sentence though:  Variable, module, function, and class names should be written between single back-ticks (`numpy`). Should I start changing the double back-ticks for this items s as I continue to make edits?