Update Variables.py
This is a first pass on a rather large file. I know the Examples section that begins at line 6 needs to be moved, but with four classes in this module, I don't know where to move it to! It also seems that some of the @properties for design properties and parent variables are duplicated.