Update Variables.py (#263)
* Update Variables.py

This is a first pass on a rather large file. I know the Examples section that begins at line 6 needs to be moved, but with four classes in this module, I don't know where to move it to! It also seems that some of the @properties for design properties and parent variables are duplicated.

* Apply suggestions from code review

Fix typos.

* Update Variables.py

Fix a syntax error.

* Update Variables.py

Fix an indentation error.

* Update Variables.py

Attempt to fix TabError: inconsistent use of tabs and spaces in indentation.

* Update Variables.py

Try again to fix the same error. Grrr.

* Update Variables.py

* Update Variables.py

* Update Variables.py

* Apply suggestions from code review

Co-authored-by: Alex Kaszynski <akascap@gmail.com>

* Update Variables.py

Co-authored-by: Alex Kaszynski <akascap@gmail.com>