Update JobManager.py (#273)
* Update JobManager.py

I don't see this module anywhere in the API doc. I added an issue about this and line 144:

def update_config_name(self, name, machinename):
"""Update the name of the machine.

I believe that line 144 is supposed to be for updating the name of the configuration file, yet it asks for the name of the machine. There is already a method for updating the machine name.

* Update pyaedt/application/JobManager.py

Fix typo.