Update Design.py
@akaszynski 
I don't understand the long description for this module. One reason is because of all of the "it" pronouns!
Also, search for "Design properties." You'll find the property "design_properties". It shows a parameter of "design_name" in the code. In the rendering, there is no list of parameters following the property name (as there is for methods). Should properties have parameters?