Update edb.py
@jleonatti 
Should edb.py refer to AEDB file? I never knew if this should be left as is or changed to AEDT!
Many of the import_ methods all say "Import a BRD file...." when another file type should be mentioned. I fixed the import_gds_file method because it was obvious, but not import_cadence_file and import_layout_pcb.