Doc/typo and very simple edits to application files (#335)
* Update desktop.py

* Update hfss.py

* Update q3d.py

* Update q3d.py

* Update edb.py

@jleonatti 
Should edb.py refer to AEDB file? I never knew if this should be left as is or changed to AEDT!
Many of the import_ methods all say "Import a BRD file...." when another file type should be mentioned. I fixed the import_gds_file method because it was obvious, but not import_cadence_file and import_layout_pcb.

* Update maxwell.py

* Update desktop.py

* Update q3d.py

* Update maxwell.py

* Update icepak.py

For create_source_power, I believe we have a typo in this new parameter: thermal_condtion.

* Update hfss3dlayout.py

* Update mechanical.py

* Create mechanical.py

* Update rmxprt.py

* Update circuit.py

Examples were duplicated in this file.

* Update simplorer.py

@jleonatti The examples refer to Maxwell projects rather than Simplorer projects. I am not sure if these are errors or not.

* Update emit.py

@jleonatti 
Examples here also say Maxwell--even though we are talking about Emit application.

* Apply suggestions from code review

Fix typos.

* Apply suggestions from code review

* Update icepak.py

* Apply suggestions from code review

Co-authored-by: jleonatti <84793532+jleonatti@users.noreply.github.com>

* Apply suggestions from code review

Co-authored-by: jleonatti <84793532+jleonatti@users.noreply.github.com>

* Apply suggestions from code review

Co-authored-by: jleonatti <84793532+jleonatti@users.noreply.github.com>

* Update hfss.py

Co-authored-by: Alex Kaszynski <akascap@gmail.com>
Co-authored-by: jleonatti <84793532+jleonatti@users.noreply.github.com>