Microdoppler updates (#393)
* Release new version 0.2.2

* Update Pipeline (#54) (#55)

* Update Pipeline

* Update Pipeline

* Apply suggestions from code review

* Merge with main 0.2.3

* Changed_Version

* Improved Edb Stability on Linux

* Fixed bug on get_report_data method to allow Near Fields Plot

* Version Update

* Padstacks

* Version and master merge

* Updated to v0.2.10

* Updated to v0.2.11

* Merged From Master
new version Tag

* Merged from Master
New Release

* Merge from Main

* Merge from Main

* Added export_hfss and export_q3d methods
Added write_export3d_option_config_file to write options file for 3d Export in Siwave

* Removed messenger object and replaced by _messenger
Updated bounding_box method

* updated version

* Added Logger Environment Variables
Improved Docs

* Bug fix on Edb
Added Ironpython UT

* Updated Dlls and fixed bug on aedt embedded Edb usage

* Improved export_q3d, export_maxwell and export_hfss methods

* Bug fixed on Maxwell2D

* Pushing 0.3.7

* Updated EdbLib.dll

* Minor documentatione edits

corrected some whitespace errors in application/variables.py

* Update CoordinateSystem.create

create with mode='axis' cannot accept x_pointing and y_pointing relative to the global CS if the expressions are defined by variables. The x_pointing and y_poining now take single arguments and are defined in the local (new) CS.

* update Primitives3D.insert_3d_component()

geoParams can have default value '' and will not
throw an error if this is not passed.

* Update GeometryModeler.create_group()

The oeditor.CreateGroup() method requires the selected objects to be passed as "SubmodelInstances:=" so they are assigned to the group passed in "Groups:=". I learned this by recording a script in 2021R2 when the command create_group() was not working as expected.

* Update version.txt

* Added archive_project method.
Fixed issues on Linux
fixed issue on units for circuit netlist creation

* Fixed Coordinate Systems method and create_group methods

* Fixed Coordinate Systems method and create_group methods

* Fixed Coordinate Systems method and create_group methods

* Fixed Coordinate Systems method and create_group methods

* Fixed Coordinate Systems method and create_group methods

* Fixed Coordinate Systems method and create_group methods

Co-authored-by: Massimo Capodiferro <massimo.capodiferro@ansys.com>
Co-authored-by: Massimo Capodiferro <77293250+maxcapodi78@users.noreply.github.com>
Co-authored-by: Alex Kaszynski <akascap@gmail.com>