There are two COMPUMAG TEAM 7 examples, one has a regressed title listing it as 2D Eddy, the second is a little better but lacks detail. Both examples use a coil geometry with square corners and no test data is included.
The new Team7 example titled Maxwell3DTeam7 has the correct coil geometry with rounded corners, it solves and then results are plotted along two lines and overlaid with the real results. This example is now consistent with the TEAM3 example and others being written.

New example does not included matplotlib lines, the plan is to add these to all the TEAM problems in more detail at a later date.