Updated TEAM7 Example to Replace Two Existing Examples (#804)
* There are two COMPUMAG TEAM 7 examples, one has a regressed title listing it as 2D Eddy, the second is a little better but lacks detail.
Both examples use a coil geometry with square corners and no test data is included.

The new Team7 example titled Maxwell3DTeam7 has the correct coil geometry with rounded corners, it solves and then results are plotted along two lines and overlaid with the real results. This example is now consistent with the TEAM3 example and others being written.

New example does not included matplotlib lines, the plan is to add these to all the TEAM problems in more detail at a later date.

* There are two COMPUMAG TEAM 7 examples, one has a regressed title listing it as 2D Eddy, the second is a little better but lacks detail.
Both examples use a coil geometry with square corners and no test data is included.

The new Team7 example titled Maxwell3DTeam7 has the correct coil geometry with rounded corners, it solves and then results are plotted along two lines and overlaid with the real results. This example is now consistent with the TEAM3 example and others being written.

New example does not included matplotlib lines, the plan is to add these to all the TEAM problems in more detail at a later date.

* There are two COMPUMAG TEAM 7 examples, one has a regressed title listing it as 2D Eddy, the second is a little better but lacks detail.
Both examples use a coil geometry with square corners and no test data is included.

The new Team7 example titled Maxwell3DTeam7 has the correct coil geometry with rounded corners, it solves and then results are plotted along two lines and overlaid with the real results. This example is now consistent with the TEAM3 example and others being written.

New example does not included matplotlib lines, the plan is to add these to all the TEAM problems in more detail at a later date.

* I've adjusted the formatting using the Black tool

* Update examples/02-Maxwell/Maxwell3DTeam7.py

Good point, it’s left over from an old method.

Co-authored-by: Maksim Beliaev <beliaev.m.s@gmail.com>

* Update examples/02-Maxwell/Maxwell3DTeam7.py

Co-authored-by: Maxime Rey <87315832+MaxJPRey@users.noreply.github.com>

* Update examples/02-Maxwell/Maxwell3DTeam7.py

Co-authored-by: Maxime Rey <87315832+MaxJPRey@users.noreply.github.com>

* Update examples/02-Maxwell/Maxwell3DTeam7.py

Change variables to lowercase

Co-authored-by: Maxime Rey <87315832+MaxJPRey@users.noreply.github.com>

* Fixed Style

Co-authored-by: Maksim Beliaev <beliaev.m.s@gmail.com>
Co-authored-by: Maxime Rey <87315832+MaxJPRey@users.noreply.github.com>
Co-authored-by: maxcapodi78 <Shark78>