Fix get_component_placement_vector (#980)
* Assume rotation about origin

Remove the usage of the merged component's center location in
get_component_placement_vector for the offset calculation.  Its return
values are expected to be passed into place_in_layout_3d_placement, and
this function applies the rotation around the origin.

* Add a test case

Add a test case for placement of an offset component on the top layer.