Fix bug for flipped mount when rotation zero
get_component_placement_vector would return an incorrect Y-offset when
flipped was True and the rotation angle was calculated to be zero.  Add
a test case for and fix this by modifying the merged component pin
locations to account for the flip around the X-axis (Y-value changes
sign) so that the correct flipped values propagate through the rest of
the calculations.