Add number_aware_string_key
Add a method to calculate keys for list.sort and sorted from strings
that treats embedded strings of digits as integers.  This is useful to
produce an "expected" ordering of component names, pin names, etc.  For
example, we expect "U2" to be sorted before "U10" in such a list but the
default sort behavior would have the opposite ordering.