Fix bug for flipped mount when rotation zero (#1009)
* Fix bug for flipped mount when rotation zero

get_component_placement_vector would return an incorrect Y-offset when
flipped was True and the rotation angle was calculated to be zero.  Add
a test case for and fix this by modifying the merged component pin
locations to account for the flip around the X-axis (Y-value changes
sign) so that the correct flipped values propagate through the rest of
the calculations.

* Feedback from review

Simplify offset calculation by not treating zero rotation as a special
case.  Also clean up the error checking for return: the existing code
implied that the vector calculation could fail when it could not and
would not capture a failure to get the solder ball height.