Doc/update edb doc (#1032)
* Update dostring for Edb.add_layer()

Add that LayerType = 1 is for dielectric.

* Update dostring for Edb.add_layer()

Added additional layer types from the EDB documentation to the docstring:
            ``2``: Conducting plane layer.
            ``3``: Airline layer.
            ``4``: Error layer.
            ``5``: Symbol layer.
            ``6``: Measure layer.
            ``8``: Assembly layer.
            ``9``: Silkscreen layer.
            ``10``: Solder Mask layer.
            ``11``: Solder Paste layer.

* Refactor example 04_edb_parameterized_design.py

The whole example was re-written based on the original version.

There are no longer class definitions in the example.
The cutout on the stripline layer allows for a smoother transition
from the via. The previous example could have the ground very close
to the signal layer at the transition.

* Add final comment to 04_edb_parameterized_design.py

This example runs as-is, but the ground planes are not
connected. I added a note at the end pointing this out.

* Update 04_edb_parametrized_design.py

var_server is not used anywhere.

Co-authored-by: Hui Zhou <ring630@gmail.com>