Failing test for configure_hfss_extents
The configure_hfss_extents method will never set its radiation_box
property because the setter checks that its argument is of type
RadiationBoxType which is just an enumeration-style class for several
constant integer values.  The setter should instead check that the
argument is an integer (and perhaps one of the allowable values).