Improved file management between client and server. (#1438)
* Improved file management between client and server.
New methods:
pyaedt_service_manager is a server listening on port 17878 and that must run on the server.
create_session the client create a new session on the server. The new session is used:
- for file transfer
- to start edb
- to start aedt

* Improved file management between client and server.
New methods:
pyaedt_service_manager is a server listening on port 17878 and that must run on the server.
create_session the client create a new session on the server. The new session is used:
- for file transfer
- to start edb
- to start aedt

* Improved file management between client and server.
New methods:
pyaedt_service_manager is a server listening on port 17878 and that must run on the server.
create_session the client create a new session on the server. The new session is used:
- for file transfer
- to start edb
- to start aedt

* improved naming

* improved naming

* Improved open files

* Improved open files

* codacy fixes

* codacy fixes

Co-authored-by: maxcapodi78 <Shark78>