Update Hfss.create_setup()
Simplify syntax and fixed a bug when MultipleAdaptiveFreqsSetup is used.  The Boolean
"MultipleAdaptiveFreqsSetup" is set to True if it is not passed (this is implied).

"AdaptMultipleFreqs" can be a list of frequencies (as strings) and the default
MaxDeltaS is used.

Also, "AdaptMultipleFreqs" can be passed as a dict with keys = freq (as it was before) but
values no longer have to be length 1 lists but can be floats.