CI: fix upload error in full documentation
Problem
Building the full documentation with sphynx leads to the generation
of a .doctree directory. This directory is composed of a file
environment.pickle which has size > 100 Mb. Thus, the upload step of
the full documentation workflow fails due to the size of such file (git
does not allow files of > 100 Mb without using lfs).

Solution
Since the .doctree directory should not be versioned at all, we had a
build-finished event to delete to .doctree directory after the
documentation build.