Feat/4010 pass primitive features (#4019)
* Pass some proprties when creating a primitive.

Update GeometryModeler._create_object in Primitives.py
Add **kwargs as last argument to most primitive creation methods in Primitives3D

* Update unit tests and documentation.

Addressed a defect #4017 that was identified while updating unit tests.
Added documentation and updated the unit tests to allow properties of primitives to be set using arguments when they are created.

* Remove comma

A comma on the last line in a list of arguments is invalid Python 2.7 syntax.

* Update argument handling for primitives

- Address a syntax change in Maxwell 2D for primitives in Primitives2D in the plane property.
- Automate assignment of any valid property to a primitive on instantiation using kwargs.

* Update test_02_2D_modeler.py

test_09_plot() was failing because the regular polygon start and first position were along the z-axis while the polygon should have been in the XY plane.  This was also the case in the "main" branch.

* Minor cleanup of comments.

* Fix error caused by transparency.

There is a defect in the native API that leads to strange behavior when transparency of 2d objects is set.

* Fix formatting to pass IronPython test.

* Resolve style conflicts

Black changes syntax that causes IronPython failures. This was resolved using # fmt: on ... # fmt: off