MAINT: use scikit-rf 0.31.0
Note: this change happens following the fix on
scikit-rf side which was trying to import a specific
function from scipy while not bounding the version
of scipy. The new version of scipy had no longer the
required function leading to CI breaks on our side
as we no longer could run the tests.