FEAT: add example documentation build check
Note: changes consist in detecting an error in the html pages context
while not stopping the overall build process (for debugging purposes).
However, a check is performed at the end of the build process
and if an error was detected, the documentation build will fail.

Warning: if this way of working changes, one would want to change
`nbsphinx_allow_errors` value and the content of the PR changes.