CI: Update workflows runs-on
Notes: to refact the CI/CD we need to avoid using pyaedt
self-hosted runners when not needed. However, the current
workflow files leverage "windows-latest" and "pyaedt" which
leads to using pyaedt self-hosted runners instead of github
runners.
On top of these changes, the self-hosted runners tag are
going to be update to remove "windows-latest" from their labels.