FIX: Deprecation warnings in docstrings
Note: Since python 3.6, backslash with something else than a
recongnized escape sequence is considered an "invalid escape
sequence". While the code was sometimes correct (others not),
it is possible that Python interpreter gets confused by the
docstring and triggers the warning because it considers the
docstring as regular strings and not raw string.