FIX: Function handler to return False
Following my request changes in #4463, it seems that the
pyaedt_function_handler function was not working correctly.
Indeed, if an exception was raised and the settings were such
that one would want the function handler to return False,
no value was returned by the function (leading to a None return).