REFACTOR: make plot methods consistent
Following #4626 and #4708, theses changes aim at making our
code base consistent. The changes consist in returning Matplotlib
figure objects on methods that were returning a list of value.
In the case of polar_plot_3d_pyvista, a Pyvista plotter object
is returned when, before, a boolean was returned if one wanted to
show the figure or save it in a file.