FIX: Handle zero-valued expression variables properly.
I noticed the Hfss.evaluate_expression() method was incorrectly
returning the symbolic value of an expression, instead of the float
value, when the expression evaluated to zero.

For example, this code would print the string 'test - test' instead of
0.0 as expected:

    app = Hfss("ZeroVariableTest")
    app['x'] = '10 mm'
    app["test"] = 'x / 1mm'

    print(f'{app.evaluate_expression('test - test')=}') # should be 0.0

I tracked it down to the Variable class constructor, where `si_value'
was being directly used as a truth value:

    if si_value:
        self._value = si_value
    else:
        self._value = self._calculated_value

Unfortunately, float(0.0) is a Falsy value in Python (`bool(0.0) ==
False`), so the above incorrectly treats a 0.0 float the same as
None. The correct condition for None-ness is `if si_value is not None`
instead of `if si_value`.

(cherry picked from commit 2e0c7530db8059881d804be6075e015599b88bd8)